% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_prob}
\alias{check_prob}
\title{check if probability vector is as expected}
\usage{
check_prob(prob)
}
\arguments{
\item{prob}{probability vector}
}
\value{
TRUE or errors
}
\description{
check if probability vector is as expected
}
