% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin.R
\name{coin}
\alias{coin}
\title{create a coin object}
\usage{
coin(object = c("heads", "tails"))
}
\arguments{
\item{object}{vector of length (both sides of the coin)}
}
\value{
a coin object
}
\description{
create a coin object
}
