\name{ExactNullDistribution-methods}
\docType{methods}
\alias{ExactNullDistribution}
\alias{ExactNullDistribution-methods}
\alias{ExactNullDistribution,ScalarIndependenceTestStatistic-method}
\alias{ExactNullDistribution,QuadTypeIndependenceTestStatistic-method}
\title{ Methods for computing exact conditional reference distributions}
\description{
 Methods for computing exact conditional reference distributions.
}
\section{Methods}{
\describe{
\item{object = "ScalarIndependenceTestStatistic"}{}
\item{object = "QuadTypeIndependenceTestStatistic"}{}
}}
\keyword{methods}
