\name{ExpectCovar-class}
\docType{class}
\alias{ExpectCovar-class}

\title{Class "ExpectCovar"}
\description{  Conditional expectation and covariance of linear statistics }
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("ExpectCovar", ...)}.
Normally, this class is used internally only.
}
\section{Slots}{
  \describe{
    \item{\code{expectation}:}{Object of class \code{"numeric"}}
    \item{\code{covariance}:}{Object of class \code{"matrix"}}
    \item{\code{dimension}:}{Object of class \code{"integer"}}
  }
}
\keyword{classes}
