% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip.coin}
\alias{flip.coin}
\title{flip a coin}
\usage{
\method{flip}{coin}(x, times = 1)
}
\arguments{
\item{x}{coin object}

\item{times}{number of times to flip the coin}
}
\value{
a vector of flips (length: times)
}
\description{
flip a coin
}
\examples{
coin1 <- coin()
flip(coin1, 5)
}
