% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{generic function for flipping an object}
\usage{
flip(x, ...)
}
\arguments{
\item{x}{object to be flipped}

\item{...}{additional parameters}
}
\value{
vector of observed elements
}
\description{
generic function for flipping an object
}
