\name{initialize-methods}
\docType{methods}
\alias{initialize}
\alias{initialize-methods}
\alias{initialize,ANY-method}
\alias{initialize,traceable-method}
\alias{initialize,signature-method}
\alias{initialize,environment-method}
\alias{initialize,ExpectCovar-method}
\alias{initialize,ExpectCovarInfluence-method}
\alias{initialize,ModelEnv-method}
\alias{initialize,CovarianceMatrix-method}
\alias{initialize,Variance-method}
\alias{initialize,IndependenceLinearStatistic-method}
\alias{initialize,IndependenceTestStatistic-method}
\title{ Methods for Function initialize in Package `coin' }
\description{
  Methods for function \code{initialize} in package \pkg{coin} 
}
\section{Methods}{
\describe{

\item{.Object = "ANY"}{}

\item{.Object = "traceable"}{}

\item{.Object = "signature"}{}

\item{.Object = "environment"}{}

\item{.Object = "ModelEnv"}{}

\item{.Object = "ExpectCovar"}{}

\item{.Object = "ExpectCovarInfluence"}{}

\item{.Object = "IndependenceProblem"}{}

\item{.Object = "IndependenceTestProblem"}{}

\item{.Object = "IndependenceTestStatistic"}{}

\item{.Object = "IndependenceLinearStatistic"}{}

\item{.Object = "ScalarIndependenceTestStatistic"}{}

\item{.Object = "MaxTypeIndependenceTestStatistic"}{}

\item{.Object = "QuadTypeIndependenceTestStatistic"}{}

\item{.Object = "SymmetryProblem"}{}
}}
\keyword{methods}
