\name{PermutationDistribtion}
\docType{methods}
\alias{pperm}
\alias{pperm-methods}
\alias{pperm,NullDistribution-method}
\alias{pperm,IndependenceTest-method}
\alias{pperm,ScalarIndependenceTest-method}
\alias{pperm,MaxTypeIndependenceTest-method}
\alias{pperm,QuadTypeIndependenceTest-method}
\alias{qperm}
\alias{qperm-methods}
\alias{qperm,NullDistribution-method}
\alias{qperm,IndependenceTest-method}
\alias{qperm,ScalarIndependenceTest-method}
\alias{qperm,MaxTypeIndependenceTest-method}
\alias{qperm,QuadTypeIndependenceTest-method}
\alias{dperm}
\alias{dperm-methods}
\alias{dperm,NullDistribution-method}
\alias{dperm,IndependenceTest-method}
\alias{dperm,ScalarIndependenceTest-method}
\alias{dperm,MaxTypeIndependenceTest-method}
\alias{dperm,QuadTypeIndependenceTest-method}
\alias{support}
\alias{support-methods}
\alias{support,NullDistribution-method}
\alias{support,IndependenceTest-method}
\alias{support,ScalarIndependenceTest-method}
\alias{support,MaxTypeIndependenceTest-method}
\alias{support,QuadTypeIndependenceTest-method}

\title{ Permutation Distribution of Conditional Independence Tests }

\description{
 Distribution, density, quantile and support of permutation distributions
 of conditional independence tests.
}
\section{Methods}{
\describe{

\item{object = "NullDistribution"}{ ~~describe this method here }

\item{object = "IndependenceTest"}{ ~~describe this method here }

\item{object = "ScalarIndependenceTest"}{ ~~describe this method here }

\item{object = "MaxTypeIndependenceTest"}{ ~~describe this method here }

\item{object = "QuadTypeIndependenceTest"}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{htest}
