% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin.R
\name{rare_coin}
\alias{rare_coin}
\title{Rare coin class}
\usage{
rare_coin(name, year, ...)
}
\arguments{
\item{name}{str, name of coin}

\item{year}{int, year coin was made}

\item{...}{parameters to feed into \code{coin}
subclass}
}
\value{
rare_coin object
}
\description{
Rare coin class
}
\examples{
rare_coin(name = "Lincoln penny", year =  1972)
}
