\name{ExactNullDistribution-methods}
\docType{methods}
\alias{ExactNullDistribution}
\alias{ExactNullDistribution-methods}
\alias{ExactNullDistribution,ScalarIndependenceTestStatistic-method}
\title{ Methods for computing exact conditional reference distributions}
\description{
 Methods for computing exact conditional reference distributions.
}
\section{Methods}{
\describe{
\item{object = "ScalarIndependenceTestStatistic"}{computes exact conditional distribution.}
}}
\keyword{methods}
