\name{PValue-class}
\docType{class}
\alias{PValue-class}

\title{Class "PValue"}
\description{ A class for p value functions }
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("PValue", ...)}.
}
\section{Slots}{
	 \describe{
    \item{\code{pvalue}:}{p value function }
    \item{\code{p}:}{distribution function }
    \item{\code{name}:}{character, name of the distribution }
  }
}
\section{Methods}{
No methods defined with class "PValue" in the signature.
}
\keyword{classes}
