\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,ANY-method}
\alias{show,traceable-method}
\alias{show,ObjectsWithPackage-method}
\alias{show,MethodDefinition-method}
\alias{show,MethodWithNext-method}
\alias{show,genericFunction-method}
\alias{show,classRepresentation-method}
\alias{show,ModelEnv-method}
\alias{show,IndependenceTest-method}
\alias{show,QuadTypeIndependenceTest-method}
\alias{show,MaxTypeIndependenceTest-method}
\alias{show,ScalarIndependenceTest-method}
\alias{show,ScalarIndependenceTestConfint-method}
\title{ Methods for Function show in Package `coin'}
\description{
  Methods for function \code{show} in package \pkg{coin}
}
\section{Methods}{
\describe{

\item{object = "IndependenceTest"}{}

\item{object = "QuadTypeIndependenceTest"}{}

\item{object = "MaxTypeIndependenceTest"}{}

\item{object = "ScalarIndependenceTest"}{}

\item{object = "ScalarIndependenceTestConfint"}{}
}}
\keyword{methods}
