\name{statistic-methods}
\docType{methods}
\alias{statistic-methods}
\alias{statistic,IndependenceTest-method}
\alias{statistic,ScalarIndependenceTestStatistic-method}
\alias{statistic,MaxTypeIndependenceTestStatistic-method}
\alias{statistic,QuadTypeIndependenceTestStatistic-method}
\title{ Extract test statistics, linear statistics and standardized
  statistics }
\description{
 Extract test statistics, linear statistics and standardized
  statistics
}
\section{Methods}{
\describe{

\item{object = "IndependenceTest"}{ ~~describe this method here }

\item{object = "ScalarIndependenceTestStatistic"}{ ~~describe this method here }

\item{object = "MaxTypeIndependenceTestStatistic"}{ ~~describe this method here }

\item{object = "QuadTypeIndependenceTestStatistic"}{ ~~describe this method here }
}}
\keyword{methods}
